backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
plac==1.1.3
preshed==3.0.2
PyJWT==1.7.1
pylint==2.4.4
pyspellchecker==0.5.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
textblob==0.15.3
thinc==7.4.0
tqdm==4.45.0
typed-ast==1.4.1
urllib3==1.25.8
wasabi==0.6.0
wrapt==1.11.2
zipp==3.1.0
